ROS2Packages/vision_model_package/ZipTilROS/requirements.txt
###### Requirements without Version Specifiers ######
opencv-python
PyYAML
open3d
numpy
ultralytics
